requirements.txt
chromadb==0.5.0
langchain==0.2.5
python-docx==1.1.2
openai==1.34.0
